Rename complaint tool as feedback tool (#8248)
* Update vets-json-schema version

* Update formId

* Update vets-json-schema version

* Update helpers, files, folders, components and other instances of complaint tool

* Update docs

* Remove obsolete ignore

* Use COMPLAINT-TOOL as formId until BE updates as well

* Nest facility code in view:searchSchoolSelect
