+++ content/pages/education/gi-bill-school-feedback.md
@@ -1,6 +1,6 @@
 ---
 title: GI Bill® School Feedback Tool
-entryname: complaint-tool
+entryname: feedback-tool
 layout: page-react.html
 description: Use your VA education benefits to pay for college or training programs. Find out which documents you’ll need to apply for benefits, and start your online application today.
 includeBreadcrumbs: true
