+++ src/applications/edu-benefits/feedback-tool/containers/FeedbackToolApp.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import RoutedSavableApp from '../../../../platform/forms/save-in-progress/RoutedSavableApp';
 import formConfig from '../config/form';
 
-export default function ComplaintToolApp({ location, children }) {
+export default function FeedbackToolApp({ location, children }) {
   return (
     <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
       {children}
