+++ src/applications/edu-benefits/feedback-tool/manifest.json
@@ -1,7 +1,7 @@
 {
     "appName": "GI Bill School Feedback Tool",
-    "entryFile": "./complaint-tool-entry.jsx",
-    "entryName": "complaint-tool",
+    "entryFile": "./feedback-tool-entry.jsx",
+    "entryName": "feedback-tool",
     "rootUrl": "/education/gi-bill-school-feedback",
     "production": false
   }
