+++ src/applications/edu-benefits/feedback-tool/reducers/schoolSelect.js
