+++ src/applications/edu-benefits/feedback-tool/selectors/schoolSearch.js
@@ -6,7 +6,7 @@ export const selectFormSubmitted = ownProps => _.get(ownProps, 'formContext.subm
 export const selectInstitutionQuery = state => _.get(state, 'schoolSelect.institutionQuery');
 export const selectInstitutions = state => _.get(state, 'schoolSelect.institutions');
 export const selectInstitutionSelected = state => _.get(state, 'schoolSelect.institutionSelected');
-export const selectManualSchoolEntryChecked = state => _.get(state, 'form.data.educationDetails.school.facilityCode.view:manualSchoolEntryChecked');
+export const selectManualSchoolEntryChecked = state => _.get(state, 'form.data.educationDetails.school.view:searchSchoolSelect.view:manualSchoolEntryChecked');
 export const selectPagesCount = state => _.get(state, 'schoolSelect.pagesCount');
 export const selectSearchInputValue = state => _.get(state, 'schoolSelect.searchInputValue');
 export const selectSearchResultsCount = state => _.get(state, 'schoolSelect.searchResultsCount');
