+++ src/applications/edu-benefits/tests/feedback-tool/00.feedback-tool.e2e.spec.js
@@ -1,7 +1,7 @@
 const E2eHelpers = require('../../../../platform/testing/e2e/helpers');
-const manifest = require('../../complaint-tool/manifest.json');
-const Timeouts = require('../../../../platform/testing/e2e/timeouts.js');
-const PageHelpers = require('./complaint-tool-helpers');
+const manifest = require('../../feedback-tool/manifest.json');
+const Timeouts = require('../../../../platform/testing/e2e/timeouts');
+const PageHelpers = require('./feedback-tool-helpers');
 const testData = require('./schema/maximal-test.json');
 const FormsTestHelpers = require('../../../../platform/testing/e2e/form-helpers');
