+++ src/applications/edu-benefits/tests/feedback-tool/actions/schoolSearch.unit.spec.js
@@ -8,7 +8,7 @@ import {
   searchSchools,
   selectInstitution,
   toggleManualSchoolEntry
-} from '../../../complaint-tool/actions/schoolSearch';
+} from '../../../feedback-tool/actions/schoolSearch';
 
 function setFetchResponse(stub, data) {
   const response = new Response(null, { headers: { 'content-type': ['application/json'] } });
