+++ src/applications/edu-benefits/tests/feedback-tool/config/contactInfo.unit.spec.js
@@ -3,11 +3,11 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import { mount } from 'enzyme';
 
-import { DefinitionTester, selectCheckbox } from '../../../../../platform/testing/unit/schemaform-utils.jsx';
-import formConfig from '../../../complaint-tool/config/form';
+import { DefinitionTester, fillData } from '../../../../../platform/testing/unit/schemaform-utils.jsx';
+import formConfig from '../../../feedback-tool/config/form';
 
-describe('complaint tool benefits info', () => {
-  const { schema, uiSchema } = formConfig.chapters.benefitsInformation.pages.benefitsInformation;
+describe('feedback tool applicant info', () => {
+  const { schema, uiSchema } = formConfig.chapters.applicantInformation.pages.contactInformation;
 
   it('should render', () => {
     const form = mount(
@@ -18,7 +18,7 @@ describe('complaint tool benefits info', () => {
         uiSchema={uiSchema}/>
     );
 
-    expect(form.find('input').length).to.equal(11);
+    expect(form.find('input').length).to.equal(7);
   });
 
   it('should not submit without required information', () => {
@@ -32,7 +32,7 @@ describe('complaint tool benefits info', () => {
     );
 
     form.find('form').simulate('submit');
-    expect(form.find('.usa-input-error').length).to.equal(1);
+    expect(form.find('.usa-input-error').length).to.equal(6);
     expect(onSubmit.called).to.be.false;
   });
 
@@ -46,7 +46,20 @@ describe('complaint tool benefits info', () => {
         uiSchema={uiSchema}/>
     );
 
-    selectCheckbox(form, 'root_educationDetails_programs_Post-9/11 Ch 33', true);
+    fillData(form, 'input#root_address_street', 'test');
+    fillData(form, 'input#root_address_street2', 'test');
+    fillData(form, 'input#root_address_city', 'test');
+    const state  = form.find('select#root_address_state');
+    state.simulate('change', {
+      target: { value: 'CA' } // TODO: update with new schema
+    });
+    const country  = form.find('select#root_address_country');
+    country.simulate('change', {
+      target: { value: 'US' }
+    });
+    fillData(form, 'input#root_address_postalCode', '12312');
+    fillData(form, 'input#root_applicantEmail', 'test@test.com');
+    fillData(form, 'input[name="root_view:applicantEmailConfirmation"]', 'test@test.com');
     form.find('form').simulate('submit');
     expect(form.find('.usa-input-error').length).to.equal(0);
     expect(onSubmit.called).to.be.true;
