+++ src/applications/edu-benefits/tests/feedback-tool/config/issueInfo.unit.spec.js
@@ -4,9 +4,9 @@ import sinon from 'sinon';
 import { mount } from 'enzyme';
 
 import { DefinitionTester, selectCheckbox, fillData } from '../../../../../platform/testing/unit/schemaform-utils.jsx';
-import formConfig from '../../../complaint-tool/config/form';
+import formConfig from '../../../feedback-tool/config/form';
 
-describe('complaint tool issue info', () => {
+describe('feedback tool issue info', () => {
   const { schema, uiSchema } = formConfig.chapters.issueInformation.pages.issueInformation;
 
   it('should render', () => {
