+++ src/applications/edu-benefits/tests/feedback-tool/reducers/schoolSelect.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import schoolSelect from '../../../complaint-tool/reducers/schoolSelect';
+import schoolSelect from '../../../feedback-tool/reducers/schoolSelect';
 
 describe('schoolSearch reducer', () => {
   describe('INSTITUTION_SELECTED', () => {
