+++ src/applications/personalization/dashboard/helpers.jsx
@@ -14,7 +14,7 @@ export const formBenefits = {
   '22-5495': 'education benefits',
   '40-10007': 'pre-need determination of eligibility in a VA national cemetery',
   VIC: 'Veteran ID Card',
-  'complaint-tool': 'feedback',
+  'COMPLAINT-TOOL': 'feedback',
   '21-686C': 'dependent status'
 };
 
@@ -46,7 +46,7 @@ export const formLinks = {
   '22-5495': '/education/apply-for-education-benefits/application/5495/',
   '40-10007': '/burials-and-memorials/pre-need/form-10007-apply-for-eligibility/',
   VIC: '/veteran-id-card/apply/',
-  'complaint-tool': '/education/complaint-tool/form',
+  'COMPLAINT-TOOL': '/education/gi-bill-school-feedback/',
   '21-686C': '/disability-benefits/686/dependent-status/'
 };
 
@@ -64,6 +64,7 @@ export const trackingPrefixes = {
   '22-5495': 'edu-5495-',
   '40-10007': 'preneed-',
   VIC: 'veteran-id-card-',
+  'COMPLAINT-TOOL': 'gi_bill_feedback',
   '21-686C': '686-'
 };
 
@@ -82,7 +83,7 @@ export const sipEnabledForms = new Set([
   '22-5495',
   '40-10007',
   'VIC',
-  'complaint-tool'
+  'COMPLAINT-TOOL'
 ]);
