+++ src/platform/forms/tests/migrations.unit.spec.js
@@ -12,7 +12,7 @@ import fullSchema527EZ from '../../../applications/pensions/config/form';
 import fullSchema530 from '../../../applications/burials/config/form';
 import fullSchema10007 from '../../../applications/pre-need/config/form';
 import fullSchema686 from '../../../applications/disability-benefits/686/config/form';
-import fullSchemaComplaintTool from '../../../applications/edu-benefits/complaint-tool/config/form';
+import fullSchemaFeedbackTool from '../../../applications/edu-benefits/feedback-tool/config/form';
 
 import schemas from 'vets-json-schema/dist/schemas';
 
@@ -30,7 +30,7 @@ const mappedIds = [
   '22-5490',
   '22-5495',
   '40-10007',
-  'complaint-tool',
+  'FEEDBACK-TOOL',
   'definitions'
 ];
 
@@ -47,7 +47,7 @@ const configs = [
   fullSchema5490,
   fullSchema5495,
   fullSchema10007,
-  fullSchemaComplaintTool
+  fullSchemaFeedbackTool
 ];
 
 const excludedForms = new Set([
@@ -64,6 +64,7 @@ describe('form migrations:', () => {
     const reformattedIds = mappedIds.slice(0);
     reformattedIds.splice(0, 1, '1010ez');
     reformattedIds.pop();
+    reformattedIds.splice(-1, 1, 'COMPLAINT-TOOL');
     const includedFormIds = configs.map(form => form.formId);
     expect(allFormIds).to.deep.equal(mappedIds);
     expect(includedFormIds).to.deep.equal(reformattedIds);
